+++ changelog.md
@@ -7,6 +7,7 @@
 * Added more options for plotting color channels for images:
   * Shortcuts `"r"`, `"g"`, and `"b"` can now be used for plotting single color channels in addition to `"red"`, `"green"`, and `"blue"`.
   * Two-channel visualizations can be plotted using `"rg"`, `"gb"`, or `"rb"`.
+* Added `duration` property to `KymoTrack` which returns the duration (in seconds) that the track was observed.
 
 #### Improvements